+++ src/main/java/com/github/zastrixarundell/toramsensei/commands/crafting/MatsCommand.java
@@ -2,6 +2,7 @@ package com.github.zastrixarundell.toramsensei.commands.crafting;
 
 import com.github.zastrixarundell.toramsensei.Parser;
 import com.github.zastrixarundell.toramsensei.commands.DiscordCommand;
+import com.github.zastrixarundell.toramsensei.commands.search.items.DiscordItemCommand;
 import com.github.zastrixarundell.toramsensei.objects.toram.Item;
 import org.javacord.api.entity.message.embed.EmbedBuilder;
 import org.javacord.api.event.message.MessageCreateEvent;
@@ -35,7 +36,7 @@ public class MatsCommand extends DiscordCommand
 
         Runnable runnable = () ->
         {
-            ArrayList<Item> justForCheck = new ArrayList<>();
+            ArrayList<Item> allItems = new ArrayList<>();
 
             try
             {
@@ -44,14 +45,9 @@ public class MatsCommand extends DiscordCommand
                         .data("special", "nalch")
                         .get();
 
-                Element table = document.getElementsByClass("table table-striped").first();
-                Element body = table.getElementsByTag("tbody").first();
+                Element cardContainer = document.getElementsByClass("card-container").first();
 
-                getItems(body).forEach(item ->
-                {
-                    sendItemEmbed(item, event);
-                    justForCheck.add(item);
-                });
+                allItems.addAll(DiscordItemCommand.getItems(cardContainer));
             }
             catch (Exception e)
             {
@@ -65,22 +61,23 @@ public class MatsCommand extends DiscordCommand
                         .data("special", "nsmith")
                         .get();
 
-                Element table = document.getElementsByClass("table table-striped").first();
-                Element body = table.getElementsByTag("tbody").first();
+                Element cardContainer = document.getElementsByClass("card-container").first();
 
-                getItems(body).forEach(item ->
-                {
-                    sendItemEmbed(item, event);
-                    justForCheck.add(item);
-                });
+                allItems.addAll(DiscordItemCommand.getItems(cardContainer));
             }
             catch (Exception e)
             {
                 e.printStackTrace();
             }
 
-            if (justForCheck.isEmpty())
+            if (allItems.isEmpty())
+            {
                 sendErrorMessage(event);
+                return;
+            }
+
+            for (int i = 0; i < allItems.size() && i < 5; i++)
+                sendItemEmbed(allItems.get(i), event);
 
         };
 
@@ -103,30 +100,15 @@ public class MatsCommand extends DiscordCommand
     private void sendErrorMessage(MessageCreateEvent messageCreateEvent)
     {
         EmbedBuilder embed = new EmbedBuilder()
-                .setTitle("Error while getting item!")
-                .setDescription("An error happened! Does the item even exist? The item may not be added yet.");
+                .setTitle("Error while getting item mats!")
+                .setDescription("An error happened! Does the specified!mats aaaa" +
+                        "item even exist? The item may not be added yet.");
 
         Parser.parseThumbnail(embed, messageCreateEvent);
         Parser.parseFooter(embed, messageCreateEvent);
         Parser.parseColor(embed, messageCreateEvent);
 
         messageCreateEvent.getChannel().sendMessage(embed);
-    }
-
-    private ArrayList<Item> getItems(Element body)
-    {
-        Elements trs = body.getElementsByTag("tr");
-
-        ArrayList<Item> listOfItems = new ArrayList<>();
-
-        for (int size = 0, count = 0; size < trs.size() && count < 5; size++)
-            if (trs.get(size).parent() == body)
-            {
-                listOfItems.add(new Item(trs.get(size)));
-                count++;
-            }
-
-        return listOfItems;
     }
 
     private void sendItemEmbed(Item item, MessageCreateEvent messageCreateEvent)